I united my tests in one file and it moved to a different location, also some fix with parse birthday from json file
