--- tests/view_information_about_new_person/__init__.py
@@ -1,1 +1,0 @@
-__author__ = 'Administrator'
